Code review comments and minor changes.
