+++ Vostok.ZooKeeper.Recipes/Helpers/DistributedLockToken.cs
@@ -62,6 +62,8 @@ namespace Vostok.ZooKeeper.Recipes.Helpers
 
                 var delete = client.DeleteProtectedAsync(new DeleteRequest(path), log).GetAwaiter().GetResult();
                 deleteResult.TrySetResult(delete);
+
+                // CR(iloktionov): EnsureSuccess()?
                 if (!delete.IsSuccessful)
                     throw new Exception("Failed to delete lock node.", delete.Exception);
 
@@ -72,6 +74,7 @@ namespace Vostok.ZooKeeper.Recipes.Helpers
             }
             else
             {
+                // CR(iloktionov): EnsureSuccess()?
                 var delete = deleteResult.Task.GetAwaiter().GetResult();
                 if (!delete.IsSuccessful)
                     throw new Exception("Failed to delete lock node.", delete.Exception);
