+++ Vostok.ZooKeeper.Recipes/Helpers/ZooKeeperStatusExtensions.cs
@@ -1,11 +1,13 @@
-﻿using Vostok.ZooKeeper.Client.Abstractions.Model;
+﻿using JetBrains.Annotations;
+using Vostok.ZooKeeper.Client.Abstractions.Model;
 using Vostok.ZooKeeper.Client.Abstractions.Model.Result;
 
 namespace Vostok.ZooKeeper.Recipes.Helpers
 {
     internal static class ZooKeeperResultExtensions
     {
-        public static bool IsRetryableError(this ZooKeeperResult result) =>
+        // CR(iloktionov): Is NotReadonlyOperation status really retriable?
+        public static bool IsRetriableError([NotNull] this ZooKeeperResult result) =>
             result.Status == ZooKeeperStatus.UnknownError ||
             result.Status == ZooKeeperStatus.NotConnected ||
             result.Status == ZooKeeperStatus.ConnectionLoss ||
